+++ DatadogInternal/Sources/NetworkInstrumentation/NetworkInstrumentationFeature.swift
@@ -103,7 +103,7 @@ internal final class NetworkInstrumentationFeature: DatadogFeature {
 
                 // Only perform interception if this swizzler should handle this task
                 // This allows the swizzler chain to continue for tasks we don't handle
-                var injectedTraceContexts: [TraceContext]?
+                var injectedTraceContexts = [RequestInstrumentationContext]()
 
                 let configuredFirstPartyHosts = FirstPartyHosts(firstPartyHosts: configuration?.firstPartyHostsTracing) ?? .init()
                 if let currentRequest = task.currentRequest {
@@ -112,7 +112,7 @@ internal final class NetworkInstrumentationFeature: DatadogFeature {
                     injectedTraceContexts = traceContexts
                 }
 
-                self.intercept(task: task, with: injectedTraceContexts ?? [], additionalFirstPartyHosts: configuredFirstPartyHosts, trackingMode: trackingMode)
+                self.intercept(task: task, with: injectedTraceContexts, additionalFirstPartyHosts: configuredFirstPartyHosts, trackingMode: trackingMode)
             }
         )
 
@@ -346,6 +346,13 @@ extension NetworkInstrumentationFeature {
         return request?.value(forHTTPHeaderField: URLRequestBuilder.HTTPHeader.ddRequestIDHeaderField) != nil
     }
 
+    /// Helper structure that optionally contains a trace context and captured state, used to pass this
+    /// information between calls.
+    struct RequestInstrumentationContext {
+        let traceContext: TraceContext?
+        let capturedState: URLSessionHandlerCapturedState?
+    }
+
     /// Intercepts the provided request by injecting trace headers based on first-party hosts configuration.
     ///
     /// Only requests with URLs that match the list of first-party hosts have tracing headers injected.
@@ -355,7 +362,7 @@ extension NetworkInstrumentationFeature {
     ///   - additionalFirstPartyHosts: Extra hosts to consider in the interception, used in conjunction with hosts defined in each handler.
     /// - Returns: A tuple containing the modified request and the list of injected TraceContexts, one or none for each handler. If no trace is injected (e.g., due to sampling),
     ///            the list will be empty.
-    internal func intercept(request: URLRequest, additionalFirstPartyHosts: FirstPartyHosts?) -> (URLRequest, [TraceContext]) {
+    internal func intercept(request: URLRequest, additionalFirstPartyHosts: FirstPartyHosts?) -> (URLRequest, [RequestInstrumentationContext]) {
         let headerTypes = firstPartyHosts(with: additionalFirstPartyHosts)
             .tracingHeaderTypes(for: request.url)
 
@@ -365,19 +372,19 @@ extension NetworkInstrumentationFeature {
 
         let networkContext = self.networkContextProvider.currentNetworkContext
         var request = request
-        var traceContexts: [TraceContext] = [] // each handler can inject distinct trace context
+
+        // TODO: RUM-13769 This code can be simplified since we never use more than one handler simultaneously.
+        var instrumentationContexts: [RequestInstrumentationContext] = [] // each handler can inject distinct instrumentation context
         for handler in handlers {
-            let (nextRequest, nextTraceContext) = handler.modify(request: request, headerTypes: headerTypes, networkContext: networkContext)
+            let (nextRequest, nextTraceContext, capturedState) = handler.modify(request: request, headerTypes: headerTypes, networkContext: networkContext)
             request = nextRequest
-            if let nextTraceContext = nextTraceContext {
-                traceContexts.append(nextTraceContext)
-            }
+            instrumentationContexts.append(.init(traceContext: nextTraceContext, capturedState: capturedState))
         }
 
         // Remove GraphQL headers before returning the modified request
         request = removeGraphQLHeadersFromRequest(request)
 
-        return (request, traceContexts)
+        return (request, instrumentationContexts)
     }
 
     /// Intercepts the provided URLSession task by creating an interception object and notifying all handlers that the interception has started.
@@ -387,7 +394,7 @@ extension NetworkInstrumentationFeature {
     ///   - injectedTraceContexts: The list of trace contexts injected into the task's request, one or none for each handler.
     ///   - additionalFirstPartyHosts: Extra hosts to consider in the interception, used in conjunction with hosts defined in each handler.
     ///   - trackingMode: The tracking mode to use for this interception (automatic or registered delegate).
-    internal func intercept(task: URLSessionTask, with injectedTraceContexts: [TraceContext], additionalFirstPartyHosts: FirstPartyHosts?, trackingMode: TrackingMode) {
+    internal func intercept(task: URLSessionTask, with instrumentationContexts: [RequestInstrumentationContext], additionalFirstPartyHosts: FirstPartyHosts?, trackingMode: TrackingMode) {
         // In response to https://github.com/DataDog/dd-sdk-ios/issues/1638 capture the current request object on the
         // caller thread and freeze its attributes through `ImmutableRequest`. This is to avoid changing the request
         // object from multiple threads:
@@ -431,7 +438,7 @@ extension NetworkInstrumentationFeature {
                 interception.register(startDate: startTime)
             }
 
-            if let traceContext = injectedTraceContexts.first {
+            if let traceContext = instrumentationContexts.compactMap({ $0.traceContext }).first {
                 // ^ If multiple trace contexts were injected (one per each handler) take the first one. This mimics the implicit
                 // behaviour from before RUM-3470.
                 interception.register(trace: traceContext)
@@ -445,9 +452,15 @@ extension NetworkInstrumentationFeature {
 
             // Only notify handlers when the interception is first created, not when it's reused
             // This prevents double notifications when both automatic and registered delegate modes are enabled
-            if isNewInterception {
-                self.handlers.forEach { $0.interceptionDidStart(interception: interception) }
-            }
+                if isNewInterception {
+                    self.handlers
+                        .forEach {
+                            $0.interceptionDidStart(
+                                interception: interception,
+                                capturedStates: instrumentationContexts.compactMap({ $0.capturedState })
+                            )
+                        }
+                }
         }
     }